+++ device_backends/include/NDRegisterAccessor.h
@@ -33,8 +33,8 @@ namespace ChimeraTK {
     /** Get or set register accessor's buffer content (1D version).
      *  @attention No bounds checking is performed, use getNumberOfSamples() to
      * obtain the number of elements in the register. */
-    UserType& accessData(unsigned int sample) { return buffer_2D[0][sample]; }
-    const UserType& accessData(unsigned int sample) const { return buffer_2D[0][sample]; }
+    UserType& accessData(size_t sample) { return buffer_2D[0][sample]; }
+    const UserType& accessData(size_t sample) const { return buffer_2D[0][sample]; }
 
     /** Get or set register accessor's buffer content (2D version).
      *  @attention No bounds checking is performed, use getNumberOfChannels() and
